changed datatype to real/float

+++ app/model.py
@@ -21,8 +21,8 @@ class Melding(Base):
 	details = Column(String)
 	laatste_activiteit = Column(Time)
 	locatie = Column(String)
-	locatie_lat = Column(Integer)
-	locatie_lon = Column(Integer)
+	locatie_lat = Column(Float)
+	locatie_lon = Column(Float)
 
 
 engine = create_engine('sqlite:///flitsers.db')
